requirements.txt
clearml
onnxoptimizer==0.2.5
cvpack2_snapdet==0.0.5.beta
dpflow
easydict
grpcio<=1.43.0
loguru>=0.5.3
numba
numpy
numpy-quaternion
nuscenes-devkit
once-devkit>=0.0.3
open3d
opencv-python-headless
pandas
pylas
pyquaternion
refile
scikit-image
scipy
spconv-cu111>=2.1.12
tensorboardX
torch==1.9.0
torchvision==0.10.0
yolox==0.2.0
